Feature deactivated and deleted user tabs (#56)
* add: deactivated users page

* add: refactor cms_users tabs and pages to avoid code duplication

* change: refactor cms_users pages, disallow list operations on deactivated users, add hasPredefinedEdit option to lists

* add: deleted users tab and extracted cms_users ajax logic to separate file

* merge: current develop and move fuction to session.php

* fix: add back disableif to deactivated users tab

* change: use listsettings helper method, remove deleted users

* change: people_trash to people_deactivated

* clean up cms_users and cms_users_deactivated
