+++ library/lib/session.php
@@ -177,3 +177,23 @@ function loadSessionData($user) {
 		FROM organisations 
 		WHERE id = :id AND (NOT organisations.deleted OR organisations.deleted IS NULL)', array('id' => $_SESSION['usergroup']['organisation_id']));
 }
+
+function loginasuser($table,$ids) {
+	$id = $ids[0];
+	if($_SESSION['user2'] or !$_SESSION['user']['is_admin']) {
+		throw new Exception('You don\'t have access. Either you are not a Boxwise God or you are already logged in as a different user!');
+	} else {
+		$_SESSION['user2'] = $_SESSION['user'];
+		$_SESSION['camp2'] = $_SESSION['camp'];
+		$_SESSION['usergroup2'] = $_SESSION['usergroup'];
+		$_SESSION['organisation2'] = $_SESSION['organisation'];
+		$_SESSION['user'] = db_row('SELECT * FROM cms_users WHERE id=:id',array('id'=>$id));
+		loadSessionData($_SESSION['user']);
+		$camplist = camplist();
+		$_SESSION['camp'] = reset($camplist);
+		$success = true;
+		$message = 'Logged in as '.$_SESSION['user']['naam'];
+	}
+
+	return array($success,$message,true);
+}
